change joincode generating to lobbyservice from entity
