+++ src/main/java/ch/uzh/ifi/hase/soprafs24/entity/Lobby.java
@@ -32,14 +32,6 @@ public class Lobby implements Serializable {
     @Column()
     private Boolean gameActive;
 
-    // Autowire LobbyRepository
-    @Autowired
-    private LobbyRepository lobbyRepository;
-
-    public Lobby() {
-        generateUniqueJoinCode(); // Generate join code upon lobby creation
-    }
-
     public Long getLobbyId() {
         return lobbyId;
     }
@@ -82,33 +74,5 @@ public class Lobby implements Serializable {
 
     public void setGameActive(Boolean gameActive) {
         this.gameActive = gameActive;
-    }
-
-    // Method to generate a unique join code
-    private void generateUniqueJoinCode() {
-        String characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-        Random rnd = new Random();
-        boolean uniqueCodeGenerated = false;
-
-        // Keep generating until a unique code is found
-        while (!uniqueCodeGenerated) {
-            StringBuilder joinCode = new StringBuilder();
-            while (joinCode.length() < 6) { // 6 characters long
-                int index = (int) (rnd.nextFloat() * characters.length());
-                joinCode.append(characters.charAt(index));
-            }
-            String potentialCode = joinCode.toString();
-            // Check if the potential code already exists in the database
-            if (!checkIfJoinCodeExists(potentialCode)) {
-                this.lobbyJoinCode = potentialCode;
-                uniqueCodeGenerated = true;
-            }
-        }
-    }
-
-    // Helper method to check if the join code already exists in the database
-    private boolean checkIfJoinCodeExists(String code) {
-        Optional<Lobby> existingLobby = lobbyRepository.findByJoinCode(code);
-        return existingLobby.isPresent();
     }
 }
