+++ src/main/java/ch/uzh/ifi/hase/soprafs24/service/LobbyService.java
@@ -14,6 +14,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Random;
 import java.util.UUID;
 
 /**
@@ -39,6 +40,37 @@ public class LobbyService {
     this.userRepository = userRepository;
   }
 
+  // Method to generate a unique join code
+  private void generateUniqueJoinCode(Lobby lobby) {
+    String characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+    Random rnd = new Random();
+    boolean uniqueCodeGenerated = false;
+
+    // Keep generating until a unique code is found
+    while (!uniqueCodeGenerated) {
+        StringBuilder joinCode = new StringBuilder();
+        while (joinCode.length() < 6) { // 6 characters long
+            int index = (int) (rnd.nextFloat() * characters.length());
+            joinCode.append(characters.charAt(index));
+        }
+        String potentialCode = joinCode.toString();
+        // Check if the potential code already exists in the database
+        if (!checkIfJoinCodeExists(potentialCode)) {
+            lobby.setLobbyJoinCode(potentialCode);
+            uniqueCodeGenerated = true;
+        }
+    }
+}
+
+// Helper method to check if the join code already exists in the database
+private boolean checkIfJoinCodeExists(String code) {
+    Lobby foundLobby = this.lobbyRepository.findByJoinCode(code).orElse(null);
+    if(foundLobby==null){
+      return false;
+    }
+    return true;
+}
+
   public List<Lobby> getLobbys() {
     return this.lobbyRepository.findAll();
   }
@@ -54,7 +86,7 @@ public class LobbyService {
 
   //TODO check after implemtning joincode in lobby entity and findbyjoincode in lobby repository(GS)
   //TODO change output from long to lobby in class diagram(GS)
-  //TODO change type of lobbyJoinCode to long(MA)
+  //TODO change type of lobbyJoinCode to long(MA) no needs to be string (GS)
   public Lobby findLobbyByJoinCode(String lobbyJoinCode) {
     Lobby foundLobby = this.lobbyRepository.findByJoinCode(lobbyJoinCode).orElse(null);
     if(foundLobby==null){
@@ -66,6 +98,8 @@ public class LobbyService {
   
   public Lobby createLobby(Long userId) {
     Lobby newLobby = new Lobby();
+
+    generateUniqueJoinCode(newLobby);
     
     newLobby.setLobbyOwner(userId);
